third_party/blink/web_tests/fast/dom/shadow/shadow-hierarchy-exception-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Appending an ancestor to a shadow tree should throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowDiv.appendChild(container) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
